ep11/requirements.txt
attrs==19.3.0
autopep8==1.5
aws-cdk.assets==1.32.0
aws-cdk.aws-apigateway==1.32.0
aws-cdk.aws-applicationautoscaling==1.32.0
aws-cdk.aws-autoscaling==1.32.0
aws-cdk.aws-autoscaling-common==1.32.0
aws-cdk.aws-autoscaling-hooktargets==1.32.0
aws-cdk.aws-batch==1.32.0
aws-cdk.aws-certificatemanager==1.32.0
aws-cdk.aws-cloudformation==1.32.0
aws-cdk.aws-cloudfront==1.32.0
aws-cdk.aws-cloudwatch==1.32.0
aws-cdk.aws-ec2==1.32.0
aws-cdk.aws-ecr==1.32.0
aws-cdk.aws-ecr-assets==1.32.0
aws-cdk.aws-ecs==1.32.0
aws-cdk.aws-eks==1.32.0
aws-cdk.aws-elasticloadbalancing==1.32.0
aws-cdk.aws-elasticloadbalancingv2==1.32.0
aws-cdk.aws-events==1.32.0
aws-cdk.aws-glue==1.32.0
aws-cdk.aws-iam==1.32.0
aws-cdk.aws-kms==1.32.0
aws-cdk.aws-lambda==1.32.0
aws-cdk.aws-logs==1.32.0
aws-cdk.aws-route53==1.32.0
aws-cdk.aws-route53-targets==1.32.0
aws-cdk.aws-s3==1.32.0
aws-cdk.aws-s3-assets==1.32.0
aws-cdk.aws-sam==1.32.0
aws-cdk.aws-secretsmanager==1.32.0
aws-cdk.aws-servicediscovery==1.32.0
aws-cdk.aws-sns==1.32.0
aws-cdk.aws-sns-subscriptions==1.32.0
aws-cdk.aws-sqs==1.32.0
aws-cdk.aws-ssm==1.32.0
aws-cdk.aws-stepfunctions==1.32.0
aws-cdk.aws-stepfunctions-tasks==1.32.0
aws-cdk.core==1.32.0
aws-cdk.custom-resources==1.32.0
aws-cdk.cx-api==1.32.0
aws-cdk.region-info==1.32.0
cattrs==1.0.0
constructs==2.0.1
jsii==1.1.0
publication==0.0.3
pycodestyle==2.5.0
python-dateutil==2.8.1
six==1.14.0
typing-extensions==3.7.4.2
